# backend/openapi.json
{
  "components": {
    "schemas": {
      "Body_import_roster_api_v1_admin_roster_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_roster_api_v1_admin_roster_post",
        "type": "object"
      },
      "Body_import_timetable_api_v1_admin_timetable_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_timetable_api_v1_admin_timetable_post",
        "type": "object"
      },
      "Body_upload_material_api_v1_materials_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_material_api_v1_materials_post",
        "type": "object"
      },
...
      "ConsentOut": {
        "properties": {
          "consent_type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          }
        },
        "required": [
          "consent_type",
          "granted",
          "recorded_at"
        ],
        "title": "ConsentOut",
        "type": "object"
      },
      "ConsentRequest": {
        "properties": {
          "consent_type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "granted": {
            "title": "Granted",
            "type": "boolean"
          }
        },
        "required": [
          "consent_type",
          "granted"
        ],
        "title": "ConsentRequest",
        "type": "object"
      },
      "ConsentType": {
        "description": "Granular. Declining one must never block the others, and declining\ncamera or microphone must never lower an engagement score.",
        "enum": [
          "terms",
          "engagement_monitoring",
          "camera",
          "microphone"
        ],
        "title": "ConsentType",
        "type": "string"
      },
...
      "Difficulty": {
        "enum": [
          "easy",
          "medium",
          "hard"
        ],
        "title": "Difficulty",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MaterialOut": {
        "properties": {
          "chunk_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Count"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/MaterialStatus"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "content_type",
          "size_bytes",
          "status",
          "uploaded_at"
        ],
        "title": "MaterialOut",
        "type": "object"
      },
      "MaterialStatus": {
        "enum": [
          "pending",
          "processing",
          "completed",
          "failed"
        ],
        "title": "MaterialStatus",
        "type": "string"
      },
      "Page_MaterialOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MaterialOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MaterialOut]",
        "type": "object"
      },
      "Page_QuestionOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[QuestionOut]",
        "type": "object"
      },
...
      "QuestionBulkReviewRequest": {
        "properties": {
          "question_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Question Ids",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          }
        },
        "required": [
          "question_ids",
          "status"
        ],
        "title": "QuestionBulkReviewRequest",
        "type": "object"
      },
      "QuestionOut": {
        "properties": {
          "correct_option": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Never sent to students. Lecturer and admin views only.",
            "title": "Correct Option"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "material_id": {
            "format": "uuid",
            "title": "Material Id",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "source_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The text the question was generated from. Grounding evidence.",
            "title": "Source Excerpt"
          },
          "source_slide": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Drives the 'see slide 3' reference in student feedback.",
            "title": "Source Slide"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType"
          }
        },
        "required": [
          "id",
          "material_id",
          "type",
          "status",
          "difficulty",
          "prompt"
        ],
        "title": "QuestionOut",
        "type": "object"
      },
      "QuestionReviewRequest": {
        "description": "A lecturer approving, editing or rejecting a generated question.",
        "properties": {
          "correct_option": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Option"
          },
          "difficulty": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Difficulty"
              },
              {
                "type": "null"
              }
            ]
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "QuestionReviewRequest",
        "type": "object"
      },
      "QuestionStatus": {
        "description": "The human-in-the-loop review workflow. A question is only ever delivered\nto students from the STAGED state, so nothing reaches a class without a\nlecturer approving it.",
        "enum": [
          "draft",
          "approved",
          "rejected",
          "staged",
          "delivered"
        ],
        "title": "QuestionStatus",
        "type": "string"
      },
      "QuestionType": {
        "enum": [
          "mcq",
          "free_text"
        ],
        "title": "QuestionType",
        "type": "string"
      },
...
      "Role": {
        "enum": [
          "student",
          "lecturer",
          "admin"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "TimetableImportResult": {
        "description": "Result of an administrator uploading a timetable CSV or XLSX.\n\nStructured data in, structured data out. Deliberately no OCR path.",
        "properties": {
          "conflicts": {
            "items": {
              "type": "string"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "rows_read": {
            "title": "Rows Read",
            "type": "integer"
          },
          "sessions_created": {
            "title": "Sessions Created",
            "type": "integer"
          },
          "unmatched_lecturers": {
            "items": {
              "type": "string"
            },
            "title": "Unmatched Lecturers",
            "type": "array"
          },
          "unmatched_students": {
            "items": {
              "type": "string"
            },
            "title": "Unmatched Students",
            "type": "array"
          }
        },
        "required": [
          "rows_read",
          "sessions_created"
        ],
        "title": "TimetableImportResult",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "token_type": {
            "const": "bearer",
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "role"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Cognitive Learning Intelligence Platform.\n\nRoutes returning 501 have a frozen contract but no implementation yet; the response body names the owning workstream.",
    "title": "C.L.I.P API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/roster": {
      "post": {
        "operationId": "import_roster_api_v1_admin_roster_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_roster_api_v1_admin_roster_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Roster",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/timetable": {
      "post": {
        "description": "CSV or XLSX only. Structured data is parsed, never OCR'd.",
        "operationId": "import_timetable_api_v1_admin_timetable_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_timetable_api_v1_admin_timetable_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableImportResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Timetable",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/auth/consent": {
      "post": {
        "description": "Consent is per-type and revocable.\n\nRevoking camera or microphone consent must stop collection immediately and\nmust not reduce the student's engagement score.",
        "operationId": "record_consent_api_v1_auth_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Consent",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/materials": {
      "get": {
        "operationId": "list_materials_api_v1_materials_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MaterialOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Materials",
        "tags": [
          "content"
        ]
      },
      "post": {
        "description": "Accepts PDF, PPTX, DOCX or TXT.\n\nReturns 202 immediately: processing runs in the background and reports\nprogress over the WebSocket as `material.progress` events.",
        "operationId": "upload_material_api_v1_materials_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_material_api_v1_materials_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Material",
        "tags": [
          "content"
        ]
      }
    },
    "/api/v1/materials/{material_id}": {
      "get": {
        "operationId": "get_material_api_v1_materials__material_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Material",
        "tags": [
          "content"
        ]
      }
    },
    "/api/v1/materials/{material_id}/questions": {
      "get": {
        "operationId": "list_questions_api_v1_materials__material_id__questions_get",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_QuestionOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Questions",
        "tags": [
          "content"
        ]
      }
    },
    "/api/v1/materials/{material_id}/questions/{question_id}": {
      "patch": {
        "description": "Approve, edit or reject a generated question.\n\nNo question reaches a student without passing through here.",
        "operationId": "review_question_api_v1_materials__material_id__questions__question_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "question_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Question Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Question",
        "tags": [
          "content"
        ]
      }
    },
    "/api/v1/materials/{material_id}/questions:bulk": {
      "post": {
        "description": "Backs the 'approve all' and 'stage N questions' actions.",
        "operationId": "bulk_review_questions_api_v1_materials__material_id__questions_bulk_post",
        "parameters": [
          {
            "in": "path",
            "name": "material_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Material Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionBulkReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionOut"
                  },
                  "title": "Response Bulk Review Questions Api V1 Materials  Material Id  Questions Bulk Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Review Questions",
        "tags": [
          "content"
        ]
      }
    },
...
  }
}